requirements.txt
requests
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
twitchio==1.1.0
websockets==8.1
yarl==1.4.2
redis==3.5.3
